apply_kernel_histogram {scattermore}	R Documentation

apply_kernel_histogram

Description

Apply a kernel to the given histogram.

Usage

apply_kernel_histogram(
  fhistogram,
  filter = "circle",
  mask = default_kernel(filter, radius, sigma),
  radius = 2,
  sigma = radius/2,
  threads = 0
)

Arguments

fhistogram
Matrix or array interpreted as histogram of floating-point values.

filter
Use the pre-defined filter, either circle, square, gauss. Defaults to circle.

mask
Custom kernel used for blurring, overrides filter. Must be a square matrix of odd size.

radius
Radius of the kernel (counted without the "middle" pixel"), defaults to 2. The generated kernel matrix will be a square with (2*radius+1) pixels on each side.

sigma
Radius of the Gaussian function selected by filter, defaults to radius/2.

threads
Number of parallel threads (default 0 chooses hardware concurrency).

Value

2D matrix with the histogram processed by the kernel application.

[Package scattermore version 1.2 Index]