scatter_lines_rgbwt {scattermore}	R Documentation

scatter_lines_rgbwt

Description

Render lines into a RGBWT bitmap.

Usage

scatter_lines_rgbwt(
  xy,
  xlim = c(min(xy[, c(1, 3)]), max(xy[, c(1, 3)])),
  ylim = c(min(xy[, c(2, 4)]), max(xy[, c(2, 4)])),
  out_size = c(512L, 512L),
  RGBA = c(0, 0, 0, 255),
  skip_start_pixel = FALSE,
  skip_end_pixel = TRUE
)

Arguments

xy
4-column matrix with point coordinates. Each row contains X and Y coordinates of line start and X and Y coordinates of line end, in this order.

xlim, ylim
2-element vector of rendered area limits (position of the first pixel on the left/top, and the last pixel on the right/bottom). You can flip the image coordinate system by flipping the *lim vectors.

out_size
2-element vector size of the result raster, defaults to c(512L,512L).

RGBA
Vector of 4 elements with integral RGBA color for the lines, defaults to c(0,0,0,255).

skip_start_pixel
TRUE if the start pixel of the lines should be omitted, defaults to FALSE.

skip_end_pixel
TRUE if the end pixel of a line should be omitted, defaults to TRUE. (When plotting long ribbons of connected lines, this prevents counting the connecting pixels twice.)

Value

Lines plotted in RGBWT bitmap.

[Package scattermore version 1.2 Index]