histogram_to_rgbwt

Colorize given histogram with input palette.

histogram_to_rgbwt(
  fhistogram,
  RGBA = grDevices::col2rgb(col, alpha = T),
  col = grDevices::hcl.colors(10),
  zlim = c(min(fhistogram), max(fhistogram))
)

Arguments

fhistogram

Matrix or 2D array with the histogram of values.

RGBA

4-by-N matrix floating-point R, G, B and A channels for the palette. Overrides col.

col

Colors to use for coloring.

zlim

Values to use as extreme values of the histogram

Value

RGBWT matrix.