scatter_points_rgbwt

Colorize given data with one given color or each point has its corresponding given color.

scatter_points_rgbwt(
  xy,
  xlim = c(min(xy[, 1]), max(xy[, 1])),
  ylim = c(min(xy[, 2]), max(xy[, 2])),
  out_size = c(512, 512),
  RGBA = c(0, 0, 0, 255),
  map = NULL,
  palette = NULL
)

Arguments

xy

2-column matrix with point coordinates. As usual with rasters in R, X axis grows right, and Y axis grows DOWN. Flipping ylim causes the "usual" mathematical behavior.

xlim

Limits as usual (position of the first pixel on the left/top, and the last pixel on the right/bottom), 2-element vector. You can easily flip the top/bottom to the "usual" mathematical system by flipping the ylim vector.

ylim

Limits as usual (position of the first pixel on the left/top, and the last pixel on the right/bottom), 2-element vector. You can easily flip the top/bottom to the "usual" mathematical system by flipping the ylim vector.

out_size

2-element vector, size of the result histogram, defaults to c(512,512).

RGBA

Vector with 4 elements or matrix or array (4xn dim, n >= 2, n ~ xy rows) with R, G, B and alpha channels in integers, defaults to c(0,0,0,255).

map

Vector with integer indices to palette.

palette

Matrix with R, G, B and A channels in rows, of at maximum(map) columns.

Value

An array in RGBWT format with the scatterplot output.